cql3: expr: add disallow_unset_value to bind_variable
Add a marker to bind_variable which disallows
passing UNSET_VALUE as a value for this bind variable.

Trying to evaluate a bind_variable that has this
marker set will cause an invalid_request_exception
if the user tries to pass UNSET_VALUE.

Having this marker will allow to make unset values
illegal in a chosen expression. In the following
commits it will be used to make them illegal
in the WHERE clause, which doesn't have any use
for unset values.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>