gossiper: do not drop seed node from the ping list
When restarting all nodes in the cluster with new IPs, the only means
to convey the new cluster topology to the nodes is by updating their
seeds. Unfortunately, this doesn't work reliably after
3b1ff90a1a80c4b0d2b6784de052a376ad75e7ee which drops the seed concept,
since if node restart takes a lot of time, the gossiper drops the node
from the peer list altogether and never attempts to contact it.

This partitions the cluster into disjoints set in rare cases.

For example, imagine there is a cluster of 4 nodes:
node1, node2, node3, node4. Node1 and node2 are used as seeds.
IP addresses of all four nodes have changed, so "hosts" table
does not contain any recognizable hosts.

The entire cluster restarts, all nodes at the same time.

It may happen that node1 and node3 will use node1 as seed and will
quickly discover each other in gossip,
node2 and node4 will use node2 as seed and quickly reach out to each
other as well.

Once a node reaches out to any non-seed node in the cluster, it stops
pinging the seeds, basically dropping the seeds from the gossiper list.
The original intent of this was to never use seeds during a restart,
i.e. use them only during initial join to the cluster. But as we can see
from the above example, the case when IP addresses of existing
cluster members change is similar but not same to the initial boot.

Anyhow, as soon as nodes in these two sub-clusters see at least one node in
"hosts" table, they drop all seeds from the ping list and hence
never attempt to contact each other.

The reason the new IPs look as new nodes during restart is
that the gossiper internally continues to identify nodes
by IP addresses, and can not identify that the node it is not being able
to reach is holding a part of the token range (only nodes holding tokens
are never dropped from the ping list).

The reason it doesn't happen during initial boot is that
DBAs are supposed to start one node at a time according to our manual
today. So in future, when we permit concurrent start, this issue
will be present during bootstrap as well. Group0 will not address
this as it will not influence gossip decision to declare a node down
and stop trying to reach it.

Fix gossiper to keep the seed nodes in the ping list as long as they
stay seeds, even if they are unreachable, and try to contact the
seeds as long as some nodes are unreachable.

In a way this patch makes a full circle from
3b1ff90a1a80c4b0d2b6784de052a376ad75e7ee

The test case follows in the next patch, and without this patch
it fails 5 times per 100 concurrent runs in debug mode.