gossiper: do not drop seed node from the ping list
When restarting all nodes in the cluster with new ips, the only means
to convey the new cluster topology to the nodes is by updating their
seeds. Unfortunately, this doesn't work reliably after
3b1ff90a1a80c4b0d2b6784de052a376ad75e7ee which drops the seed concept,
since if node restart takes a lot of time, the gossiper drops the node
from the peer list altogether and never attempts to contact it.

This partitions the cluster into disjoints set in rare cases.

This happens because the gossiper internally continues to identify nodes
by ip addresses, and can not identify that the node it is not being able
to reach is holding a part of the token range (for such nodes, never are
never dropped from the ping list), and assumes it's a node which failied
to join the cluster.

Fix gossiper to keep the seed nodes in the ping list as long as they
stay seeds, even if they are unreachable.