test: refresh_nodes() in wait_for_cql_and_get_hosts()
Wait for CQL and get hosts is an utility function
used by the test to determine that the cluster is fully up.

It fetches cluster metadata from the control connection
and verifies that there is a connection pool to every node in the
cluster. If a node is still down, there could be no connection pool.

The way the driver watches DOWN -> UP transitions is push
notifications from the control node over control connection:
when starting a control connection, it reads the hosts system
table, and then, any change to the hosts system table is expected
to provide a push notification.

Oh well, it doesn't. The procedure of fetching the hosts table,
updating the local peers database on the driver, marking the
node up and sending a push notificaiton is inherently race prone.
A push notification may arrive ahead of fully reading the hosts
table, so a node is removed from the list *after* it has been
marked up. Running the tests for 100 times concurrently in debug mode
shows the race

So instead of relying on a push notification explicitly refresh the node
list.