test/sstable::test_env: Keep track of auto-incrementing generation
Lots of test cases make sstables with monotonically incrementing
generation values. In Scylla code this counter is maintained in class
table, but sstable tests not always have it. To mimic this behavior, the
test_env can keep track of the generation, so that callers just don't
mess with it (next patch).

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>