replica: table: add make_new_generation function
make_new_generation generates a new generation
from an optional one.

If disengaged, it just generates a new generation
based on the shard_id.  Otherwise, it generates
the next generation in sequence by adding
smp::count to the previous value, like we do today.

In the future, with uuid-based generations, the
function could be used to generate a new random
uuid based on the optional parameter.

It will be up to the caller, e.g. replica::table or
sstables manager to decide which kind of generation to
create.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>