migration_manager: expose access to storage_proxy
A migration_manager holds a reference to a storage_proxy, and uses it
internally a lot - e.g., to gain access to the data_dictionary. Users
of migration_manager might also benefit from this storage_proxy - we
will see such a case in the next patches. So let's provide a getter
for the storage_proxy.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>